+++ changelog.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.14.0] - 2025-03-05
+## [0.14.0] - 2025-03-06
 ### Added
-- Add a new method `getStatus`
+- Add a new launch option `cache` for caching loaded resources
+- Add a new method getStatus
+- `loadState` now supports loading resolvable files
 
 ## [0.13.0] - 2025-02-28
 ### Added
@@ -153,6 +155,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix core loading ([9790327](https://github.com/arianrhodsandlot/nostalgist/commit/97903270c39788b3d05b7f7d0700ccdf32503a9e))
 
 [Unreleased]: https://github.com/arianrhodsandlot/eslint-config/compare/v0.14.0...HEAD
+[0.14.0]: https://github.com/arianrhodsandlot/eslint-config/compare/v0.14.0...v0.14.0
 [0.14.0]: https://github.com/arianrhodsandlot/eslint-config/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/arianrhodsandlot/eslint-config/compare/v0.12.1...v0.13.0
 [0.12.1]: https://github.com/arianrhodsandlot/eslint-config/compare/v0.12.0...v0.12.1